[FE: feat]: 리버스 지오코드 성공함 이제 정상적으로 시/군 시/군/구 표시됨 본문: 현재위치 행정구역 받아올 수 있고, 드롭다운에서 지역 선택하면 지역 코드로 변환해줌

+++ fe/D104/app/src/main/java/com/d104/pnt/data/remote/model/request/ChatRequest.kt
@@ -8,7 +8,7 @@ data class ChatCreateRequest(
     @SerializedName("description")
     val description: String,
     @SerializedName("regionCode")
-    val regionCode: Long,
+    val regionCode: Int,
     @SerializedName("maxMembers")
     val maxMembers: Int
 )
